refactor: 2d: move shared viewmodelbuilders to a neutral home (#1423)
