--- app/viewModelBuilders/catalog/brc-analytics-catalog/common/constants.ts
@@ -1,5 +1,0 @@
-export const GALAXY_DATACACHE = "Galaxy Data Cache";
-export const GENOME_BROWSER = "UCSC Genome Browser";
-export const NCBI_ASSEMBLY = "NCBI Genome Assembly";
-export const NCBI_TAXONOMY = "NCBI Taxonomy";
-export const NCBI_DATASETS_URL = "https://www.ncbi.nlm.nih.gov/datasets";
